API ApiCommentController refactor

+++ helloscala-backend/server/src/main/java/com/helloscala/admin/controller/view/BOCategoryArticleCountView.java
@@ -11,6 +11,5 @@ public class BOCategoryArticleCountView {
 
     private String name;
 
-
-    private Long count;
+    private Integer count;
 }
